+++ packages/core/src/http-middlewares/before/add-query-params.js
@@ -14,7 +14,24 @@ const addQueryParams = (req) => {
 
     normalizeEmptyParamFields(req);
 
-    const stringifiedParams = querystring.stringify(req.params);
+    let stringifiedParams = querystring.stringify(req.params);
+
+    // it goes against spec, but for compatibility, some APIs want certain
+    // characters (mostly $) unencoded
+    if (req.skipEncodingChars) {
+      for (let i = 0; i < req.skipEncodingChars.length; i++) {
+        const char = req.skipEncodingChars.charAt(i);
+        const valToReplace = querystring.escape(char);
+        if (valToReplace === char) {
+          continue;
+        }
+        // no replaceAll in JS yet, coming in a node version soon!
+        stringifiedParams = stringifiedParams.replace(
+          new RegExp(valToReplace, 'g'),
+          char
+        );
+      }
+    }
 
     if (stringifiedParams) {
       req.url += `${splitter}${stringifiedParams}`;