+++ packages/core/src/tools/create-app-request-client.js
@@ -16,9 +16,10 @@ const oauth1SignRequest = require('../http-middlewares/before/oauth1-sign-reques
 const prepareRequest = require('../http-middlewares/before/prepare-request');
 
 // after middles
-const logResponse = require('../http-middlewares/after/log-response');
+const { logResponse } = require('../http-middlewares/after/log-response');
 const prepareResponse = require('../http-middlewares/after/prepare-response');
-const throwForStatus = require('../http-middlewares/after/throw-for-status');
+const throwForStaleAuth = require('../http-middlewares/after/throw-for-stale-auth');
+const throwForStatusMiddleware = require('../http-middlewares/after/throw-for-status');
 
 const createAppRequestClient = (input, options) => {
   input = ensurePath(input, '_zapier.app');
@@ -51,11 +52,22 @@ const createAppRequestClient = (input, options) => {
     httpBefores.push(disableSSLCertCheck);
   }
 
+  let includeAutoRefresh = false;
+  if (
+    app.authentication &&
+    (app.authentication.type === 'session' ||
+      (app.authentication.type === 'oauth2' &&
+        _.get(app, 'authentication.oauth2Config.autoRefresh')))
+  ) {
+    includeAutoRefresh = true;
+  }
+
   const httpAfters = [
     prepareResponse,
     logResponse,
+    ...(includeAutoRefresh ? [throwForStaleAuth] : []),
     ...ensureArray(app.afterResponse),
-    throwForStatus,
+    throwForStatusMiddleware,
   ];
 
   return createRequestClient(httpBefores, httpAfters, options);