+++ packages/core/test/create-app.js
@@ -320,111 +320,119 @@ describe('create-app', () => {
   });
 
   describe('HTTP after middleware for auth refresh', () => {
-    it('should be applied to OAuth2 refresh app on shorthand requests', (done) => {
-      const oauth2AppDefinition = dataTools.deepCopy(appDefinition);
-      oauth2AppDefinition.authentication = {
-        type: 'oauth2',
-        test: {},
-        oauth2Config: {
-          authorizeUrl: {}, // stub, not needed for this test
-          getAccessToken: {}, // stub, not needed for this test
-          autoRefresh: true,
-        },
-      };
-      const oauth2App = createApp(oauth2AppDefinition);
+    // general purpose response tester
+    const testResponse = async (appDef, useShorthand, errorVerifier) => {
+      const methodPart = useShorthand
+        ? 'executeRequestAsShorthand'
+        : 'executeRequestAsFunc';
 
+      const app = createApp(appDef);
       const event = {
         command: 'execute',
         bundle: {
-          inputData: {
-            url: `${HTTPBIN_URL}/status/401`,
-          },
+          inputData: useShorthand
+            ? {
+                url: `${HTTPBIN_URL}/status/401`,
+              }
+            : {
+                options: {
+                  url: `${HTTPBIN_URL}/status/401`,
+                },
+              },
         },
-        method: 'resources.executeRequestAsShorthand.list.operation.perform',
+        method: `resources.${methodPart}.list.operation.perform`,
       };
-      oauth2App(createInput(oauth2AppDefinition, event, testLogger))
-        .then(() => {
-          done('expected an error, got success');
-        })
-        .catch((error) => {
-          should(error).instanceOf(errors.ResponseError);
-          error.name.should.eql('ResponseError');
-          const response = JSON.parse(error.message);
-          should(response.status).equal(401);
-          done();
-        });
-    });
 
-    it('should be applied to OAuth2 refresh app on z.request in functions', (done) => {
-      const oauth2AppDefinition = dataTools.deepCopy(appDefinition);
-      oauth2AppDefinition.authentication = {
-        type: 'oauth2',
-        test: {},
-        oauth2Config: {
-          authorizeUrl: {}, // stub, not needed for this test
-          getAccessToken: {}, // stub, not needed for this test
-          autoRefresh: true,
-        },
-      };
-      const oauth2App = createApp(oauth2AppDefinition);
+      const input = createInput(appDef, event, testLogger);
+      const err = await app(input).should.be.rejected();
 
-      const event = {
-        command: 'execute',
-        bundle: {
-          inputData: {
-            options: {
-              url: `${HTTPBIN_URL}/status/401`,
+      errorVerifier(err);
+    };
+    // the two types of error verification
+    const verifyRefreshAuthError = (error) => {
+      should(error).instanceOf(errors.RefreshAuthError);
+      error.name.should.eql('RefreshAuthError');
+    };
+
+    const verifyResponseError = (error) => {
+      should(error).instanceOf(errors.ResponseError);
+      error.name.should.eql('ResponseError');
+      const response = JSON.parse(error.message);
+      should(response.status).equal(401);
+    };
+
+    it('should be applied to OAuth2 refresh app on shorthand requests', () => {
+      return testResponse(
+        {
+          ...appDefinition,
+          authentication: {
+            type: 'oauth2',
+            oauth2Config: {
+              autoRefresh: true,
             },
           },
         },
-        method: 'resources.executeRequestAsFunc.list.operation.perform',
-      };
-      oauth2App(createInput(oauth2AppDefinition, event, testLogger))
-        .then(() => {
-          done('expected an error, got success');
-        })
-        .catch((error) => {
-          should(error).instanceOf(errors.ResponseError);
-          error.name.should.eql('ResponseError');
-          const response = JSON.parse(error.message);
-          should(response.status).equal(401);
-          done();
-        });
+        true,
+        verifyRefreshAuthError
+      );
     });
 
-    it('should be applied to session auth app on z.request in functions', (done) => {
-      const sessionAuthAppDefinition = dataTools.deepCopy(appDefinition);
-      sessionAuthAppDefinition.authentication = {
-        type: 'session',
-        test: {},
-        sessionConfig: {
-          perform: {}, // stub, not needed for this test
-        },
-      };
-      const sessionAuthApp = createApp(sessionAuthAppDefinition);
-
-      const event = {
-        command: 'execute',
-        bundle: {
-          inputData: {
-            options: {
-              url: `${HTTPBIN_URL}/status/401`,
+    it('should be applied to OAuth2 refresh app on z.request in functions', () => {
+      return testResponse(
+        {
+          ...appDefinition,
+          authentication: {
+            type: 'oauth2',
+            oauth2Config: {
+              autoRefresh: true,
             },
           },
         },
-        method: 'resources.executeRequestAsFunc.list.operation.perform',
-      };
-      sessionAuthApp(createInput(sessionAuthAppDefinition, event, testLogger))
-        .then(() => {
-          done('expected an error, got success');
-        })
-        .catch((error) => {
-          should(error).instanceOf(errors.ResponseError);
-          error.name.should.eql('ResponseError');
-          const response = JSON.parse(error.message);
-          should(response.status).equal(401);
-          done();
-        });
+        false,
+        verifyRefreshAuthError
+      );
+    });
+
+    it('should not be applied to OAuth2 refresh app on z.request in functions by default', () => {
+      return testResponse(
+        {
+          ...appDefinition,
+          authentication: {
+            type: 'oauth2',
+            oauth2Config: {
+              autoRefresh: false,
+            },
+          },
+        },
+        false,
+        verifyResponseError
+      );
+    });
+
+    it('should be applied to session auth app on z.request in functions', () => {
+      return testResponse(
+        {
+          ...appDefinition,
+          authentication: {
+            type: 'session',
+          },
+        },
+        false,
+        verifyRefreshAuthError
+      );
+    });
+
+    it('should not be applied to custom auth app on z.request in functions', () => {
+      return testResponse(
+        {
+          ...appDefinition,
+          authentication: {
+            type: 'custom',
+          },
+        },
+        false,
+        verifyResponseError
+      );
     });
   });
 
@@ -655,6 +663,184 @@ describe('create-app', () => {
       ).should.be.rejectedWith(
         /Response needs to be JSON, form-urlencoded or parsed in middleware/
       );
+    });
+  });
+
+  describe('error skipping', () => {
+    describe('shorthand requests', () => {
+      const method =
+        'resources.executeRequestAsShorthand.create.operation.perform';
+      it('should throw on shorthand by default', async () => {
+        const app = createApp(appDefinition);
+
+        const event = {
+          command: 'execute',
+          bundle: {
+            inputData: {
+              url: `${HTTPBIN_URL}/status/400`,
+            },
+          },
+          method,
+        };
+        const err = await app(
+          createInput(appDefinition, event, testLogger)
+        ).should.be.rejected();
+        JSON.parse(err.message).status.should.eql(400);
+      });
+
+      it('should throw on shorthand even when global skip flag is present', async () => {
+        const appDef = {
+          ...appDefinition,
+          flags: {
+            // this is ignored for shorthand requests
+            skipThrowForStatus: true,
+          },
+        };
+        const app = createApp(appDef);
+
+        const event = {
+          command: 'execute',
+          bundle: {
+            inputData: {
+              url: `${HTTPBIN_URL}/status/400`,
+            },
+          },
+          method,
+        };
+        const err = await app(
+          createInput(appDef, event, testLogger)
+        ).should.be.rejected();
+        JSON.parse(err.message).status.should.eql(400);
+      });
+
+      it('should skip throw on shorthand if local flag is set', async () => {
+        const appDef = {
+          ...appDefinition,
+        };
+        appDef.resources.executeRequestAsShorthand.create.operation.perform.skipThrowForStatus = true;
+
+        const app = createApp(appDef);
+
+        // httpbin can't return a body w/ the status endpoint
+        const url = 'https://status-with-body.example.com';
+        nock(url)
+          .get('/')
+          .reply(400, { status: 400, message: 'mocked response' });
+
+        const event = {
+          command: 'execute',
+          bundle: { inputData: { url } },
+          method,
+        };
+        const { results } = await app(
+          createInput(appDef, event, testLogger)
+        ).should.be.fulfilled();
+
+        results.status.should.eql(400);
+        results.message.should.eql('mocked response');
+      });
+    });
+
+    describe('z.request requests', () => {
+      const method = 'resources.executeRequestAsFunc.list.operation.perform';
+      it('should throw on request by default', async () => {
+        const app = createApp(appDefinition);
+
+        const event = {
+          command: 'execute',
+          bundle: {
+            inputData: {
+              options: { url: `${HTTPBIN_URL}/status/400` },
+            },
+          },
+          method,
+        };
+        const err = await app(
+          createInput(appDefinition, event, testLogger)
+        ).should.be.rejected();
+
+        JSON.parse(err.message).status.should.eql(400);
+      });
+
+      it('should skip throw when global skip flag is present', async () => {
+        const appDef = {
+          ...appDefinition,
+          flags: {
+            skipThrowForStatus: true,
+          },
+        };
+        // httpbin can't return a body w/ the status endpoint
+        const url = 'https://status-with-body.example.com';
+        nock(url)
+          .get('/')
+          .reply(400, [{ id: 1, status: 400, message: 'mocked response' }]);
+
+        const app = createApp(appDef);
+        const event = {
+          command: 'execute',
+          bundle: {
+            inputData: { options: { url } },
+          },
+          method,
+        };
+        const { results } = await app(
+          createInput(appDef, event, testLogger)
+        ).should.be.fulfilled();
+
+        results[0].status.should.eql(400);
+        results[0].message.should.eql('mocked response');
+      });
+
+      it('should skip throw if local flag is set', async () => {
+        const app = createApp(appDefinition);
+
+        // httpbin can't return a body w/ the status endpoint
+        const url = 'https://status-with-body.example.com';
+        nock(url)
+          .get('/')
+          .reply(400, [{ id: 1, status: 400, message: 'mocked response' }]);
+
+        const event = {
+          command: 'execute',
+          bundle: {
+            inputData: { options: { url, skipThrowForStatus: true } },
+          },
+          method,
+        };
+        const { results } = await app(
+          createInput(appDefinition, event, testLogger)
+        ).should.be.fulfilled();
+
+        results[0].status.should.eql(400);
+        results[0].message.should.eql('mocked response');
+      });
+
+      it('should throw if local flag is false and global flag is set', async () => {
+        const appDef = {
+          ...appDefinition,
+          flags: {
+            skipThrowForStatus: true,
+          },
+        };
+        const app = createApp(appDef);
+
+        const event = {
+          command: 'execute',
+          bundle: {
+            inputData: {
+              options: {
+                url: `${HTTPBIN_URL}/status/400`,
+                skipThrowForStatus: false,
+              },
+            },
+          },
+          method,
+        };
+        const err = await app(
+          createInput(appDefinition, event, testLogger)
+        ).should.be.rejected();
+        JSON.parse(err.message).status.should.eql(400);
+      });
     });
   });
 });