+++ packages/legacy-scripting-runner/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 3.8.7
+
+- :bug: Replace deasync with synckit to fix hanging ([#509](https://github.com/zapier/zapier-platform/pull/509))
+
+## 3.8.6
+
+- :bug: Fix issues with file uploading ([#496](https://github.com/zapier/zapier-platform/pull/496))
+- :nail_care: Pass `legacy.skipEncodingChars` to `z.request()` ([#501](https://github.com/zapier/zapier-platform/pull/501))
+
 ## 3.8.5
 
 - :bug: Prune nulls and undefined's from query params ([#446](https://github.com/zapier/zapier-platform/pull/446))